Flyttet filtermeny til egen komponent
