+++ apps/skde/pages/sykehusprofil/index.tsx
@@ -1,9 +1,4 @@
-import React, { useState, useEffect } from "react";
-import {
-  useQueryParam,
-  DelimitedArrayParam,
-  withDefault,
-} from "use-query-params";
+import React, { useState } from "react";
 import { UseQueryResult } from "@tanstack/react-query";
 import {
   Header,
@@ -12,32 +7,17 @@ import {
 } from "../../src/components/Header";
 import {
   skdeTheme,
-  FilterSettingsValue,
-  FilterMenu,
   useUnitNamesQuery,
-  useUnitUrlsQuery,
   defaultYear,
-  TreeViewFilterSection,
-  getTreatmentUnitsTree,
-  FilterSettings,
-  CustomAccordionExpandIcon,
   mainHospitals,
 } from "qmongjs";
 import { Footer } from "../../src/components/Footer";
-import {
-  ThemeProvider,
-  Box,
-  Accordion,
-  AccordionSummary,
-  AccordionDetails,
-  Container,
-  styled,
-} from "@mui/material";
+import { ThemeProvider, Box, Container } from "@mui/material";
 import Grid from "@mui/material/Grid2";
-import { ClickAwayListener } from "@mui/base";
+
 import { PageWrapper } from "../../src/components/StyledComponents/PageWrapper";
 import { HospitalInfoBox } from "../../src/components/HospitalProfile";
-import { URLs } from "types";
+
 import { getUnitFullName } from "../../src/helpers/functions/getUnitFullName";
 import { AffiliatedHospitals } from "../../src/components/HospitalProfile/AffiliatedHospitals";
 import { useScreenSize } from "@visx/responsive";
@@ -45,25 +25,12 @@ import { breakpoints } from "qmongjs";
 import { HospitalProfileMedfieldTable } from "../../src/components/HospitalProfile/HospitalProfileMedfieldTable";
 import { HospitalProfileLowLevelTable } from "../../src/components/HospitalProfile/HospitalProfileLowLevelTable";
 import { HospitalProfileLinePlot } from "../../src/components/HospitalProfile/HospitalProfileLinePlot";
-
-const AccordionWrapper = styled(Box)(() => ({
-  "& MuiAccordion-root:before": {
-    backgroundColor: "white",
-  },
-}));
+import { UnitFilterMenu } from "../../src/components/HospitalProfile/UnitFilterMenu";
 
 export const Skde = (): JSX.Element => {
-  const [expanded, setExpanded] = useState(false);
-
-  const treatmentUnitsKey = "selected_treatment_units";
+  const [unitName, setUnitName] = useState<string>();
 
   const { width } = useScreenSize();
-
-  // Current unit name and its setter function
-  const [selectedTreatmentUnits, setSelectedTreatmentUnits] = useQueryParam(
-    treatmentUnitsKey,
-    withDefault(DelimitedArrayParam, ["Nasjonalt"]),
-  );
 
   // Infobox URL
   const [unitUrl, setUnitUrl] = useState<string | null>(null);
@@ -91,76 +58,10 @@ export const Skde = (): JSX.Element => {
 
     unitFullName =
       unitNamesQuery.data &&
-      getUnitFullName(
-        unitNamesQuery.data.nestedUnitNames,
-        selectedTreatmentUnits[0],
-      );
-  }
-
-  const treatmentUnits = getTreatmentUnitsTree(unitNamesQuery);
-
-  if (treatmentUnits.treedata.length > 1) {
-    // Find the index of "Private" and remove the children. The sub units should not be shown.
-    // TreetmentUnits.treedata starts with one element "Nasjonalt". Need to wait for it to build up the rest.
-    const indPrivate = treatmentUnits.treedata.findIndex(
-      (x) => x.nodeValue.value === "Private",
-    );
-    treatmentUnits.treedata[indPrivate].children = [];
+      getUnitFullName(unitNamesQuery.data.nestedUnitNames, unitName);
   }
 
   // The following code ensures that the page renders correctly
-  const unitUrlsQuery = useUnitUrlsQuery();
-
-  const [mounted, setMounted] = useState(false);
-
-  useEffect(() => {
-    setMounted(true);
-  }, []);
-
-  const shouldRefreshInitialState = mounted && unitUrlsQuery.isFetched;
-
-  // Callback function for initialising the filter meny
-  const initialiseFilter = (
-    filterInput: Map<string, FilterSettingsValue[]>,
-  ) => {
-    const newUnit = filterInput.get(treatmentUnitsKey).map((el) => el.value);
-
-    let unitUrl: URLs | undefined;
-    if (unitUrlsQuery.data) {
-      unitUrl = unitUrlsQuery.data.filter((row: URLs) => {
-        return row.shortName === newUnit[0];
-      });
-    }
-
-    if (unitUrl && unitUrl[0]) {
-      setUnitUrl(unitUrl[0].url);
-    } else {
-      setUnitUrl(null);
-    }
-  };
-
-  // Callback function for updating the filter menu
-  const handleChange = (filterInput: FilterSettings) => {
-    const newUnit = filterInput.map
-      .get(treatmentUnitsKey)
-      .map((el) => el.value);
-
-    setExpanded(false);
-    setSelectedTreatmentUnits(newUnit);
-
-    let unitUrl: URLs | undefined;
-    if (unitUrlsQuery.data) {
-      unitUrl = unitUrlsQuery.data.filter((row: URLs) => {
-        return row.shortName === newUnit[0];
-      });
-    }
-
-    if (unitUrl && unitUrl[0]) {
-      setUnitUrl(unitUrl[0].url);
-    } else {
-      setUnitUrl(null);
-    }
-  };
 
   // Year for filtering
   const lastYear = defaultYear;
@@ -202,60 +103,12 @@ export const Skde = (): JSX.Element => {
           breadcrumbs={breadcrumbs}
           maxWidth={maxWidth}
         >
-          <ClickAwayListener onClickAway={() => setExpanded(false)}>
-            <AccordionWrapper>
-              <Accordion
-                square={true}
-                sx={{
-                  width: Math.min(400, 0.8 * width),
-                  borderRadius: 11,
-                  border: 1,
-                  borderColor: skdeTheme.palette.primary.main,
-                  backgroundColor: "white",
-                  color: skdeTheme.palette.primary.main,
-                }}
-                expanded={expanded}
-                onChange={(e, expanded) => {
-                  setExpanded(expanded);
-                }}
-              >
-                <AccordionSummary expandIcon={<CustomAccordionExpandIcon />}>
-                  <h3>
-                    {selectedTreatmentUnits[0] === "Nasjonalt"
-                      ? "Velg behandlingssted"
-                      : selectedTreatmentUnits[0]}
-                  </h3>
-                </AccordionSummary>
-
-                <AccordionDetails>
-                  <FilterMenu
-                    refreshState={shouldRefreshInitialState}
-                    onSelectionChanged={handleChange}
-                    onFilterInitialized={initialiseFilter}
-                  >
-                    <TreeViewFilterSection
-                      refreshState={shouldRefreshInitialState}
-                      treedata={treatmentUnits.treedata}
-                      defaultvalues={treatmentUnits.defaults}
-                      initialselections={
-                        selectedTreatmentUnits.map((value) => ({
-                          value: value,
-                          valueLabel: value,
-                        })) as FilterSettingsValue[]
-                      }
-                      sectionid={treatmentUnitsKey}
-                      sectiontitle={"Behandlingsenheter"}
-                      filterkey={treatmentUnitsKey}
-                      searchbox={true}
-                      multiselect={false}
-                      accordion={false}
-                      noShadow={true}
-                    />
-                  </FilterMenu>
-                </AccordionDetails>
-              </Accordion>
-            </AccordionWrapper>
-          </ClickAwayListener>
+          <UnitFilterMenu
+            width={Math.min(400, 0.8 * width)}
+            setUnitName={setUnitName}
+            setUnitUrl={setUnitUrl}
+            unitNamesQuery={unitNamesQuery}
+          />
         </Header>
 
         <Container maxWidth={maxWidth} disableGutters={true}>
@@ -265,7 +118,7 @@ export const Skde = (): JSX.Element => {
                 <HospitalInfoBox
                   boxHeight={width > breakpoints.xxl ? 350 : 450}
                   unitNames={unitNamesQuery.data}
-                  selectedTreatmentUnit={selectedTreatmentUnits[0]}
+                  selectedTreatmentUnit={unitName}
                   unitUrl={unitUrl}
                 />
               </Grid>
@@ -274,7 +127,7 @@ export const Skde = (): JSX.Element => {
                   boxHeight={width > breakpoints.xxl ? 350 : 450}
                   titleStyle={titleStyle}
                   unitNames={unitNamesQuery.data}
-                  selectedTreatmentUnit={selectedTreatmentUnits[0]}
+                  selectedTreatmentUnit={unitName}
                 />
               </Grid>
 
@@ -284,14 +137,14 @@ export const Skde = (): JSX.Element => {
                   titlePadding={titlePadding}
                   titleStyle={titleStyle}
                   textMargin={textMargin}
-                  unitName={selectedTreatmentUnits[0]}
+                  unitName={unitName}
                   lastYear={lastYear}
                 />
               </Grid>
 
               <Grid size={{ xs: 12 }}>
                 <HospitalProfileLowLevelTable
-                  unitName={selectedTreatmentUnits[0]}
+                  unitName={unitName}
                   boxMaxHeight={boxMaxHeight}
                   titlePadding={titlePadding}
                   titleStyle={titleStyle}
@@ -304,7 +157,7 @@ export const Skde = (): JSX.Element => {
               <Grid size={{ xs: 12 }}>
                 <HospitalProfileLinePlot
                   unitFullName={unitFullName}
-                  unitNames={selectedTreatmentUnits[0]}
+                  unitNames={unitName}
                   lastYear={lastYear}
                   pastYears={pastYears}
                   titlePadding={titlePadding}
